% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnum-package.R
\docType{package}
\name{cnum-package}
\alias{cnum}
\alias{cnum-package}
\title{cnum: Working with Chinese Numerals}
\description{
This R package provides useful functions to work with Chinese numerals in R,
such as conversion between Chinese numerals and Arabic numerals as well as
detection and extraction of Chinese numerals in character objects and string.
}
\note{
Due to technical limitation, R package documentation cannot contain any
  non-ASCII characters. Therefore, Chinese characters are represented in
  romanized Chinese \emph{pinyin} in the documentation. Visit the GitHub page
  for examples in Chinese.
}
\section{Warnings}{
 This package supports conversion of numbers with absolute
  value not greater than 1e+18. Note that numbers in R are in double
  precision that carries approximately 16 significant digits. The conversion
  accuracy for numbers beyond this limit is therefore not guaranteed.
}

\seealso{
GitHub page: \url{https://github.com/elgarteo/cnum}
}
\author{
Elgar Teo (\email{elgarteo@connect.hku.hk})
}
\keyword{internal}
