% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-lang.R
\name{default_cnum_lang}
\alias{default_cnum_lang}
\title{Default Language for \code{cnum}}
\usage{
default_cnum_lang()
}
\value{
The default language for \code{cnum} functions.
}
\description{
Function to check the default language for \code{cnum} functions.
}
\details{
This package supports Traditional Chinese and Simplified Chinese. The
language can be specified with the \code{lang} parameter in every function,
with \code{"tc"} for Traditional Chinese and \code{"sc"} for Simplified
Chinese. The default is \code{"tc"}, but this can be changed by setting
\code{\link[base]{options}(cnum.lang = "sc")}.
}
\examples{
\donttest{
# Set the default language to Simplified Chinese
options(cnum.lang = "sc")
}
default_cnum_lang()

}
\seealso{
\itemize{ \item \link[=conversion]{Functions for conversion} \item
  \link[=tools]{Functions for detetction and extraction} }
}
