% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnum-package.R
\docType{package}
\name{cnum-package}
\alias{cnum-package}
\alias{cnum}
\title{cnum: Working with Chinese Numerals}
\description{
This R package provides useful functions to work with Chinese numerals in R,
such as conversion between Chinese numerals and Arabic numerals as well as
detection and extraction of Chinese numerals in character objects and string.
}
\note{
Due to technical limitation, R package documentation cannot contain
  Chinese characters. Therefore, the Chinese characters are represented in
  romainized Chinese \emph{pinyin} in the documentation and substituted with
  \emph{"hello"} in some of the example code. Visit the GitHub page for
  clearer examples with Chinese characters.
}
\section{Warnings}{
 The precision of converting large numbers and long
  decimals is limited in base R. You might be able to improve the accurancy
  by running \code{options(digits = 22)}.
}

\seealso{
GitHub page: \url{https://github.com/elgarteo/cnum}
}
\author{
Elgar Teo (\email{elgarteo@connect.hku.hk})
}
